Move DEBUG code paths from the main SceneDelegate to a new `DebuggingSceneDelegate` subclass to separate debug- and test-specific code from production code
